Add Castle Windsor container unit tests for the Controller Installer
Prior to this commit, the only way to validate if a container had all of the necessary registrations
was to upload the resource files and manually test it. This method is very error prone, wasted a
lot of developer time, and sometimes a defect could be sitting in the registration and developers
wouldn't know until it randomly came up in testing / production.

It will also be noticed that _all_ of the registration has been moved to the ControllerInstaller. This
is on purpose. For more information as to _why_ we're moving towards a "Composition Root" as
opposed to a composition by project, please see the following:
http://blog.ploeh.dk/2011/07/28/CompositionRoot/