Maked old installers as obsolete and fixed issue with web's castle windsor validation test
We are moving the codebase to use a single Composition Root per application. This means
that we will no longer be placing IWindsorInstaller files in each project, but rather, at each
entry point of the application. Thus, the Web and Agent applications now have a single
Composition Root but the event handlers and Kepler service still need to be updated.

Finally, it should be noted that all installers that are not Composition Roots have been marked
as obsolete and should no longer be used. There is a considerable amount of code still using
these installers and for SOI and time, we will be removing these usages at a later time.