[REL-329985] Update SingleObjectFieldSanitizer to account for Object Manager export behavior
When an exported Single Object field is unset, Object Manager will return a JSON object
with ArtifactID = 0 and no Name instead of a null value (like it does for Single Choice).
This commit updates the relevant sanitizer to check for an ArtifactID of zero, and if so
it returns null for the new value.